Create multifigure module and move figure and table modules into their own folder
