+++ src/ssb_dash_framework/assets/stylesheet.css
@@ -197,6 +197,16 @@ Figuredisplay
     width: 100%;
 }
 
+.multifigure {
+    height: 100%;
+    width: 100%;
+}
+
+.multifigure-content {
+    height: 100%;
+    width: 100%;
+}
+
 /*
 SkjemapdfViewer
 */
