+++ src/ssb_dash_framework/modules/__init__.py
@@ -1,12 +1,13 @@
-"""Modules for use in the application, directly or implmented as a view (tab/window)."""
+"""Modules for use in the application, implmented as a view (tab/window) or directly with a custom layout implementation."""
 
 from .aarsregnskap import Aarsregnskap
 from .bofregistry import BofInformation
-from .figuredisplay import FigureDisplay
+from .building_blocks import EditingTable
+from .building_blocks import FigureDisplay
+from .building_blocks import MultiFigure
+from .building_blocks import MultiTable
 from .freesearch import FreeSearch
 from .skjemapdfviewer import SkjemapdfViewer
-from .tables import EditingTable
-from .tables import MultiTable
 
 __all__ = [
     "Aarsregnskap",
@@ -14,6 +15,7 @@ __all__ = [
     "EditingTable",
     "FigureDisplay",
     "FreeSearch",
+    "MultiFigure",
     "MultiTable",
     "SkjemapdfViewer",
 ]
