+++ src/ssb_dash_framework/windows/__init__.py
@@ -2,6 +2,7 @@
 
 from .bofregistry_window import BofInformationWindow
 from .figuredisplay_window import FigureDisplayWindow
+from .figuredisplay_window import MultiFigureWindow
 from .freesearch_window import FreeSearchWindow
 from .skjemapdfviewer_window import SkjemapdfViewerWindow
 from .tables_window import EditingTableWindow
@@ -12,6 +13,7 @@ __all__ = [
     "EditingTableWindow",
     "FigureDisplayWindow",
     "FreeSearchWindow",
+    "MultiFigureWindow",
     "MultiTableWindow",
     "SkjemapdfViewerWindow",
 ]
