+++ src/ssb_dash_framework/windows/figuredisplay_window.py
@@ -1,4 +1,5 @@
-from ..modules.figuredisplay import FigureDisplay
+from ..modules.building_blocks.figuredisplay import FigureDisplay
+from ..modules.building_blocks.figuredisplay import MultiFigure
 from ..utils import WindowImplementation
 
 
@@ -12,5 +13,15 @@ class FigureDisplayWindow(WindowImplementation, FigureDisplay):
             figure_func=figure_func,
             output=output,
             clickdata_func=clickdata_func,
+        )
+        WindowImplementation.__init__(self)
+
+
+class MultiFigureWindow(WindowImplementation, MultiFigure):
+    def __init__(self, label, figure_list):
+        MultiFigure.__init__(
+            self,
+            label=label,
+            figure_list=figure_list,
         )
         WindowImplementation.__init__(self)
